refactor(chat): extract group-realism map helpers to a part file
Move the low-level private helpers that read/write the per-character group
realism map (`_getCurrentSpeakerIdForRealism`, `_setGroupRealismValue`,
`_getGroupInt`, `_getGroupString`, `_coerceNeedsVector`, `_getGroupNeeds`,
`_setGroupNeeds`) into chat_service_group_realism_helpers.dart as a `part of`
extension. They are pure plumbing over `_groupRealism` (no engine logic, no
orchestration) and, being private, are never part of the public interface — so
moving them to an extension is safe (no `implements`/fake can shadow them).

Verbatim move, zero behaviour change. chat_service.dart 9,993 → 9,931 lines.
flutter analyze clean; full suite (1428 unit + golden) green.
